Create `TalpidVpnService` super class
